Change test based on the model update
