+++ src/brouwers/kitreviews/templates/kitreviews/base.html
@@ -1,6 +1,7 @@
 {% extends "base.html" %}
 {% load i18n staticfiles %}
 
+
 {% block style %}
     {{ block.super }}
     <link rel="stylesheet" type="text/css" href="{% static 'css/kitreviews.css' %}" charset="utf-8"/>
