+++ src/brouwers/kitreviews/templates/kitreviews/review_form_field_old.html
@@ -1,0 +1,5 @@
+<label for="{{ field.id_for_label }}"{% if field.field.required %} class="required"{% endif %}>
+    {{ field.label }}
+</label>
+{{ field }}
+{% include "field_errors.html" %}
