+++ src/brouwers/kitreviews/templates/kitreviews/review_form_old.html
@@ -1,0 +1,58 @@
+{% load i18n %}
+{% load sniplates %}
+
+<div id="review_kit">
+    {% form_field form.model_kit %}
+    {% include "kitreviews/review_form_field.html" %}
+
+</div>
+
+<div id="review_raw_text">
+    {% with reviewform.raw_text as field %}
+        {% include "kitreviews/review_form_field.html" %}
+    {% endwith %}
+</div>
+
+<div id="pos-neg-sides">
+    <div id="review_pos_points">
+        {% with reviewform.positive_points as field %}
+            {% include "kitreviews/review_form_field.html" %}
+        {% endwith %}
+    </div>
+    <div id="review_neg_points">
+        {% with reviewform.negative_points as field %}
+            {% include "kitreviews/review_form_field.html" %}
+        {% endwith %}
+    </div>
+</div>
+
+<div id="review_rating">
+    {% with reviewform.rating as field %}
+        {% include "kitreviews/review_form_field.html" %}
+    {% endwith %}
+</div>
+
+<div class="tab midpage">{% trans "Album, build report(s)" %}</div>
+
+<div id="review_links">
+    <p>
+        {% with reviewform.album as field %}
+            {% include "kitreviews/review_form_field.html" %}
+        {% endwith %}
+    </p>
+    <p>
+        {% with reviewform.topic_id as field %}
+            {% include "kitreviews/review_form_field.html" %}
+        {% endwith %}
+    </p>
+    <p>
+        {% with reviewform.external_topic_url as field %}
+            {% include "kitreviews/review_form_field.html" %}
+        {% endwith %}
+    </p>
+    <p>
+        {% with reviewform.show_real_name as field %}
+            {% include "kitreviews/review_form_field.html" %}
+        {% endwith %}
+    </p>
+</div>
